pyproject.toml
[project]
name = "flaskr"
dependencies = [
    "flask",
    "Flask-SQLAlchemy",
    "flask-serialize"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"